+++ front/src/components/auth/Signup.vue
@@ -111,7 +111,7 @@
 </template>
 
 <script>
-import store from '@/store';
+import store from '../../store';
 
 export default {
   name: 'Signup',