+++ front/src/components/core/Toolbar.vue
@@ -34,7 +34,7 @@
 </template>
 
 <script>
-import store from '@/store';
+import store from '../../store';
 
 export default {
   data: () => ({