+++ front/src/router/dashboard.js
@@ -2,7 +2,7 @@ export default [
   {
     path: '/dashboard',
     redirect: '/dashboard/overview',
-    component: () => import('@/views/Dashboard.vue'),
+    component: () => import('../views/Dashboard.vue'),
     children: [
       {
         path: 'overview',
@@ -11,7 +11,7 @@ export default [
           requiresAuth: true,
           displayName: 'Overview',
         },
-        component: () => import('@/components/dashboard/Overview.vue'),
+        component: () => import('../components/dashboard/Overview.vue'),
       },
       {
         path: 'root',
@@ -20,7 +20,7 @@ export default [
           requiresAuth: true,
           displayName: 'Root',
         },
-        component: () => import('@/components/dashboard/Root.vue'),
+        component: () => import('../components/dashboard/Root.vue'),
       },
       {
         path: 'intermediate/:id',
@@ -29,7 +29,7 @@ export default [
           requiresAuth: true,
           displayName: 'Intermediate',
         },
-        component: () => import('@/components/dashboard/Intermediate'),
+        component: () => import('../components/dashboard/Intermediate.vue'),
       },
       {
         path: 'certificate/:id',
@@ -38,7 +38,7 @@ export default [
           requiresAuth: true,
           displayName: 'Certificate',
         },
-        component: () => import('@/components/dashboard/Certificate.vue'),
+        component: () => import('../components/dashboard/Certificate.vue'),
       },
       {
         path: 'user/profile',
@@ -47,7 +47,7 @@ export default [
           requiresAuth: true,
           displayName: 'Profile',
         },
-        component: () => import('@/components/dashboard/user/UserProfile'),
+        component: () => import('../components/dashboard/user/UserProfile.vue'),
       },
     ],
   },